Authorization implemented

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
+import { User } from 'src/auth/user.entity';
 
 // Services are like modules which can be injected in any place and be used
 // They are singletones which means there is only 1 instance of service at the time
@@ -22,12 +23,14 @@ export class TasksService {
   // only be able to manipulate thru method in this service
   // const tasks: string[] = []
 
-  async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-    return this.taskRepository.getTasks(filterDto);
+  async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+    return this.taskRepository.getTasks(filterDto, user);
   }
 
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const found = await this.taskRepository.findOne({
+      where: { id, userId: user.id },
+    });
 
     // Throws it here, and NestJs on the background will recognize it as Http expection and return response to the client with formatted error
     if (!found) {
@@ -36,22 +39,26 @@ export class TasksService {
     return found;
   }
 
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDto);
+  async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user);
   }
 
-  async deleteTask(id: number): Promise<void> {
+  async deleteTask(id: number, user: User): Promise<void> {
     // remove means that you have to firstly retreive the entity from db and then pass entity as an argument for remove method
     // which is quite expensive, while delete - deletes entity by criteria and does not require any extra work to be done
-    const result = await this.taskRepository.delete(id);
+    const result = await this.taskRepository.delete({ id, userId: user.id });
     // amount of rows affected by query
     if (result.affected === 0) {
       throw new NotFoundException(`Could not find requested task (ID: ${id})`);
     }
   }
 
-  async updateStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateStatus(
+    id: number,
+    status: TaskStatus,
+    user: User,
+  ): Promise<Task> {
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
     return task;
